chore: remove uploader.js and move all into index.js
